requirements.txt
torch== 1.13.1
torchvision==0.14.1
fastai==2.7.12
Flask
Flask-Cors
pillow